+++ CHANGELOG.md
@@ -13,11 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Export of html preview file. [#24](https://github.com/clowder-framework/CONSORT-frontend/issues/24)
 - All client actions are recorded in a space. [#29](https://github.com/clowder-framework/CONSORT-frontend/issues/29)
 - Client can handle pdf file uploads and triggers appropriate extractors. [#31](https://github.com/clowder-framework/CONSORT-frontend/issues/31)
+
 ### Changed
 
 
 ### Fixed
-
+- Display of error messages on uploading files in dropzone. [#26](https://github.com/clowder-framework/CONSORT-frontend/issues/26)
 
 ### Security